ciao-down: Reorganize the source
The code involved in managing the VM once it has been booted has been moved
to a new file, vm.go.  All the remaining code in cloudinit.go has been moved
to prepare.go.  Thus the code that deals with setting up the VM is now in
prepare.go.  The code that manages the running vm is in vm.go.  cloudinit.go
still exists, but contains only the templates.
